requirements.txt
aiocache==0.12.3
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiosignal==1.3.2
alpaca-py==0.39.3
amdsmi @ file:///opt/rocm-6.4.0/share/amd_smi
annotated-types==0.7.0
anyio==4.9.0
appdirs==1.4.4
argcomplete==3.1.4
asgiref==3.8.1
asyncio==3.4.3
asyncpg==0.30.0
attrs==23.2.0
Automat==22.10.0
Babel==2.10.3
bcc==0.29.1
bcrypt==3.2.2
beautifulsoup4==4.12.3
blinker==1.9.0
boto3==1.34.46
botocore==1.34.46
bottle==0.12.25
Brotli==1.1.0
build==1.2.2.post1
CacheControl==0.14.2
certifi==2023.11.17
chardet==5.2.0
cleo==2.1.0
click==8.1.6
cloud-init==24.4.1
colorama==0.4.6
command-not-found==0.3
configobj==5.0.8
constantly==23.10.4
contourpy==1.0.7
coverage==7.8.0
crashtest==0.4.1
cryptography==41.0.7
cssselect==1.2.0
cycler==0.11.0
dbus-python==1.3.2
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.18
distlib==0.3.9
distro==1.9.0
distro-info==1.7+build1
dnspython==2.6.1
docker==5.0.3
docker-compose==1.29.2
dockerpty==0.4.1
docopt==0.6.2
dulwich==0.22.8
fastapi==0.115.12
fastjsonschema==2.21.1
filelock==3.18.0
findpython==0.6.3
Flask==3.1.0
fonttools==4.46.0
frozenlist==1.5.0
fs==2.4.16
Glances==3.4.0.3
googleapis-common-protos==1.69.2
gpg==1.18.0
greenlet==3.1.1
grpcio==1.71.0
gyp==0.1
h11==0.14.0
hidapi==0.9.0.post3
hiredis==3.1.0
html5lib==1.1
httpcore==1.0.7
httplib2==0.20.4
httpx==0.28.1
hyperlink==21.0.0
idna==3.6
importlib_metadata==8.6.1
incremental==22.10.0
influxdb==5.3.1
iniconfig==2.1.0
installer==0.7.0
iotop==0.6
itsdangerous==2.2.0
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.1.0
jeepney==0.9.0
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.4.2
jsonpatch==1.32
jsonpointer==2.0
jsonschema==4.10.3
keyring==25.6.0
kiwisolver==1.4.8
launchpadlib==1.11.0
lazr.restfulclient==0.14.6
lazr.uri==1.0.6
lxml==5.2.1
lz4==4.0.2+dfsg
Markdown==3.5.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.10.1
mdurl==0.1.2
more-itertools==10.6.0
motor==3.7.0
mpmath==0.0.0
msgpack==1.1.0
multidict==6.2.0
netaddr==0.8.0
netifaces==0.11.0
numpy==1.26.4
oauthlib==3.2.2
olefile==0.46
opentelemetry-api==1.31.1
opentelemetry-exporter-otlp==1.31.1
opentelemetry-exporter-otlp-proto-common==1.31.1
opentelemetry-exporter-otlp-proto-grpc==1.31.1
opentelemetry-exporter-otlp-proto-http==1.31.1
opentelemetry-instrumentation==0.52b1
opentelemetry-instrumentation-aiohttp-client==0.52b1
opentelemetry-instrumentation-asgi==0.52b1
opentelemetry-instrumentation-fastapi==0.52b1
opentelemetry-instrumentation-redis==0.52b1
opentelemetry-instrumentation-sqlalchemy==0.52b1
opentelemetry-proto==1.31.1
opentelemetry-sdk==1.31.1
opentelemetry-semantic-conventions==0.52b1
opentelemetry-util-http==0.52b1
orjson==3.10.16
packaging==24.0
pandas==2.2.3
pbs-installer==2025.3.17
pexpect==4.9.0
pillow==10.2.0
pipx==1.4.3
pkginfo==1.12.1.2
platformdirs==4.2.0
pluggy==1.5.0
ply==3.11
poetry==2.1.2
poetry-core==2.1.2
prometheus_client==0.21.1
propcache==0.3.1
protobuf==5.29.4
psutil==5.9.8
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycryptodomex==3.20.0
pydantic==2.11.3
pydantic_core==2.33.1
Pygments==2.17.2
PyGObject==3.48.2
PyHamcrest==2.1.0
PyJWT==2.7.0
pymongo==4.12.0
pyOpenSSL==23.2.0
pyparsing==3.1.1
pyproject_hooks==1.2.0
pyrsistent==0.20.0
pyserial==3.5
pysmi==0.3.4
pysnmp==4.4.12
pystache==0.6.0
pytest==8.3.5
pytest-asyncio==0.26.0
pytest-cov==6.1.1
python-apt==2.7.7+ubuntu4
python-dateutil==2.8.2
python-debian==0.1.49+ubuntu2
python-dotenv==1.0.1
python-magic==0.4.27
pytz==2024.1
PyYAML==6.0.1
pyzmq==26.4.0
RapidFuzz==3.13.0
ratelimit==2.2.1
redis==5.2.1
requests==2.31.0
requests-toolbelt==1.0.0
rich==13.7.1
s3transfer==0.10.1
scikit-learn==1.6.1
scipy==1.15.2
SecretStorage==3.3.3
service-identity==24.1.0
setuptools==68.1.2
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
sos==4.8.2
soupsieve==2.5
SQLAlchemy==2.0.40
ssh-import-id==5.11
starlette==0.46.1
sympy==1.12
systemd-python==235
texttable==1.6.7
threadpoolctl==3.6.0
tomlkit==0.13.2
trove-classifiers==2025.3.19.19
Twisted==24.3.0
typing-inspection==0.4.0
typing_extensions==4.13.1
tzdata==2025.2
ubuntu-drivers-common==0.0.0
ubuntu-pro-client==8001
ufoLib2==0.16.0
ufw==0.36.2
ujson==5.9.0
unattended-upgrades==0.1
unicodedata2==15.1.0
urllib3==2.0.7
userpath==1.9.1
uvicorn==0.27.1
uvloop==0.19.0
virtualenv==20.30.0
wadllib==1.3.6
webencodings==0.5.1
websocket-client==1.7.0
Werkzeug==3.1.3
wheel==0.42.0
wrapt==1.17.2
wsproto==1.2.0
xkit==0.0.0
yarl==1.19.0
yfinance==0.2.55
zipp==3.21.0
zope.interface==6.1
zstandard==0.23.0